refactor: reubicados archivos en `assets/`
