+++ frontend/src/components/EstablecimientoCard/EstablecimientoCard.tsx
@@ -12,7 +12,7 @@ import { PhoneIcon } from "@chakra-ui/icons";
 import { Establecimiento } from "@/models/index";
 import { Link } from "react-router-dom";
 import { Box } from "@chakra-ui/react";
-import { DEFAULT_IMAGE_SRC } from "@/utils/consts";
+import { FALLBACK_IMAGE_SRC } from "@/utils/consts";
 
 type EstablecimientoCardProps = {
   establecimiento: Establecimiento;
@@ -35,7 +35,7 @@ export default function EstablecimientoCard({
             src={
               !(establecimiento?.urlImagen === null)
                 ? establecimiento?.urlImagen
-                : DEFAULT_IMAGE_SRC
+                : FALLBACK_IMAGE_SRC
             }
             borderTopRadius="lg"
             alt={`Imagen del establecimiento ${establecimiento.nombre}`}
