+++ frontend/src/pages/_app.tsx
@@ -2,7 +2,7 @@ import { Box, Heading, Image, Text } from "@chakra-ui/react";
 import { Outlet } from "react-router-dom";
 import TopMenu from "@/components/TopMenu/TopMenu";
 import { Modals } from "@generouted/react-router";
-import ErrorPageIllustrationSvg from "@/assets/svg/error_page_illustration.svg";
+import ErrorPageIllustrationSvg from "@/assets/error_page_illustration.svg";
 
 // `Catch` es tratado por el generouted plugin, si hay algún error al cargar la página.
 export const Catch = () => {
